BC BREAK - namespace/structure refactoring, code format
